+++ lib/pio.rb
@@ -1,17 +1,9 @@
 # -*- coding: utf-8 -*-
+
+require 'pio/parse_error'
+
 require 'pio/arp'
 require 'pio/lldp'
 require 'pio/icmp'
 require 'pio/mac'
 require 'pio/dhcp'
-
-module Pio
-  # Raised when the packet data is in wrong format.
-  class ParseError < StandardError; end
-end
-
-### Local variables:
-### mode: Ruby
-### coding: utf-8-unix
-### indent-tabs-mode: nil
-### End: