+++ lib/pio/icmp.rb
@@ -4,17 +4,12 @@ require 'pio/icmp/reply'
 require 'pio/icmp/request'
 require 'pio/message_type_selector'
 
+# Packet parser and generator library.
 module Pio
   # Icmp parser and generator.
   class Icmp
     extend MessageTypeSelector
-
     message_type Request::TYPE => Request, Reply::TYPE => Reply
   end
+  ICMP = Icmp
 end
-
-### Local variables:
-### mode: Ruby
-### coding: utf-8-unix
-### indent-tabs-mode: nil
-### End: