+++ lib/pio/parse_error.rb
@@ -1,0 +1,4 @@
+module Pio
+  # Raised when the packet data is in wrong format.
+  class ParseError < StandardError; end
+end