+++ spec/spec_helper.rb
@@ -26,9 +26,3 @@ if ENV['TRAVIS']
   require 'coveralls'
   Coveralls.wear!
 end
-
-### Local variables:
-### mode: Ruby
-### coding: utf-8-unix
-### indent-tabs-mode: nil
-### End: